# Optimize settings storage and delegation