In my testing this improves minimum start up time (i.e. in steady state after sbt has cached everything from a previous run, testing with `time sbt exit && time sbt exit && time sbt exit` using the last result) for cinnamon by ~10 seconds both on 1.1.0 and 1.1.1.

Startup time 1.1.0 Vanilla: 56s
With the patch: 44s
Startup time 1.1.1 Vanilla: 63s
With the patch: 52s (together with the validId patch 45s)

This is quite a big improvement for such a self-contained change (and if you know what sbt else is doing startup...)

It also show that sbt 1.1.1 seems to have an unrelated performance regression that costs ~ 5s.

/cc @bantonsson