@jrudolph 2018-03-06
Thanks for the comments, @eed3si9n. Just FYI, I'm not sure when I'm able to continue on this, maybe end of the week.