@jrudolph 2019-08-27
> I tested this patch with the current development head (after resolving conflicts). I found the speedups to be a bit more modest in the akka project. It took about 18.75 seconds to load with the current development HEAD and 18.25 seconds with this patch applied.

Yeah, that doesn't seem to be enough alone to warrant that change. I guess it would need some reprofiling at this point before deciding if it is still worth something or what needs to be done to improve it to get the speedups I measured back then.